test: fold image resolution tests into asset-imports suite
`repro-image.test.ts` was a scratch name for tests that belong with the
existing `Content Layer - Asset Imports` suite. Move both cases there and
rename the local `schema` helper to `imageSchema` now that it shares a
module.
